clean up encoder tests into a single file
